% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdbconpkg.R
\name{close}
\alias{close}
\title{Close connection to Q}
\usage{
close(manager, handle)
}
\arguments{
\item{manager}{reference to Java object in memory}

\item{handle}{int}
}
\value{
nothing
}
\description{
Close connection to Q
}

