% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdbconpkg.R
\name{connect}
\alias{connect}
\title{Open connection to Q}
\usage{
connect(manager, host, port)
}
\arguments{
\item{manager}{reference to Java object in memory}

\item{host}{string}

\item{port}{string or 5000L as numeric}
}
\value{
int handle to Q connection
}
\description{
Open connection to Q
}

