% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdbconpkg.R
\name{exec}
\alias{exec}
\title{Execute Q statement remotely
Conversion (KDB)98h ->(Java)KxTable->(R)data frame}
\usage{
exec(manager, handle, query)
}
\arguments{
\item{manager}{reference to java class}

\item{handle}{int connection handle}

\item{query}{Q string}
}
\value{
nothing
}
\description{
Execute Q statement remotely
Conversion (KDB)98h ->(Java)KxTable->(R)data frame
}

