% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdbconpkg.R
\name{select}
\alias{select}
\title{Select Q table as R data frame
Conversion (KDB)98h ->(Java)KxTable->(R)data frame}
\usage{
select(manager, handle, query)
}
\arguments{
\item{manager}{reference to java class}

\item{handle}{int connection handle}

\item{query}{Q string}
}
\value{
data frame
}
\description{
Select Q table as R data frame
Conversion (KDB)98h ->(Java)KxTable->(R)data frame
}

